data-pipeline/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
pandas==2.0.0
google-cloud-bigquery==3.10.0
google-cloud-storage==2.8.0
